Add serdes module for string args
This factors some of the COM bus protocol serialization and
deserialization code out of betrusted-ec and xous-core. The point
is to reduce the potential for mis-matches between how the payloads
are encoded vs. decoded.

Changes:
1. Add serdes::StringSer and serdes::StringDes structs using const
   generics to support serializing and deserializing different
   sizes of strings as COM bus verb payloads.
2. Add tests to validate round trip encoding and decoding.

Credits: Thanks to xobs for suggestions on using iterators for more
efficient array access compared to C-style indexing.

Signed-off-by: samblenny